requirements.txt
python-telegram-bot
requests
python-dotenv
asyncio
duckduckgo-search

# Benchmark system requirements
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.11.0  # For statistical analysis
scikit-learn>=1.3.0  # For ML-based optimization

# Visualization
plotly>=5.18.0
kaleido  # For static image export
nbformat  # For notebook integration

# Statistics and ML
statsmodels>=0.14.0  # For advanced statistical analysis
optuna>=3.4.0  # For hyperparameter optimization
bayesian-optimization>=1.4.0  # For Bayesian optimization
lightgbm>=4.1.0  # For gradient boosting models
xgboost>=2.0.0  # For gradient boosting models

# Data processing
joblib>=1.3.0  # For parallel processing
tqdm>=4.66.0  # For progress bars
pyyaml>=6.0.1  # For configuration files
jsonschema>=4.20.0  # For JSON validation

# Development tools
pytest>=7.4.0  # For testing
black>=23.12.0  # For code formatting
pylint>=3.0.0  # For code analysis
mypy>=1.8.0  # For type checking

# Documentation
mkdocs>=1.5.0  # For documentation generation
mkdocs-material>=9.5.0  # For documentation theme
pdoc>=14.1.0  # For API documentation
